Move verifyToken to middleware
